+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ dev
 - Added the MaxVar acquisition method
 - Added the RandMaxVar acquisition method
 - Fix elfi.Distance to support scipy 1.0.0
+- Added the ExpIntVar acquisition method
 
 0.6.3 (2017-09-28)
 ------------------